"merge commit"

+++ G3000/g3000.h
@@ -22,8 +22,8 @@ public:
 
 public slots:
     bool GENERATORS_EXPORT turnOn(bool i_on) Q_DECL_OVERRIDE;
-    bool GENERATORS_EXPORT setAmp(float &m_amp) Q_DECL_OVERRIDE;
-    bool GENERATORS_EXPORT setFrequency(float &m_f) Q_DECL_OVERRIDE;
+    bool GENERATORS_EXPORT setAmp(double &m_amp) Q_DECL_OVERRIDE;
+    bool GENERATORS_EXPORT setFrequency(double &m_f) Q_DECL_OVERRIDE;
 
 public:
     void GENERATORS_EXPORT setFrequencyGrid(int i_frequencyGrid) Q_DECL_OVERRIDE;
@@ -38,19 +38,19 @@ public:
 signals:
 //    void error(QString e);
 //    void disconnected();
-//    void newFrequency(float freq_Hz);
-//    void newAmplitude(float amp_V);
-//    void newTFm(float t_s);
+//    void newFrequency(double freq_Hz);
+//    void newAmplitude(double amp_V);
+//    void newTFm(double t_s);
 
 protected:
 
     void initialize();
-    bool setAttenuation(float &attenuation);
-    float getAttenuation();
+    bool setAttenuation(double &attenuation);
+    double getAttenuation();
 
     bool commute(quint8);
     bool checkResponse();
-    float getReferenceFrequency(int refFreq);
+    double getReferenceFrequency(int refFreq);
 
 private:
 
@@ -67,11 +67,11 @@ private:
         DiffSignal
     };
 
-    float referenceFrequency;
+    double referenceFrequency;
 
-    float attenuationMax;
-    float attenuationMin;
-    float attenuationStep;
+    double attenuationMax;
+    double attenuationMin;
+    double attenuationStep;
 
     Head3000 head;
     Response3000 response;
