+++ getu.h
@@ -14,8 +14,8 @@ public:
 
 public slots:
     bool GENERATORS_EXPORT turnOn(bool i_on) Q_DECL_OVERRIDE;
-    bool GENERATORS_EXPORT setAmp(float &m_amp) Q_DECL_OVERRIDE;
-    bool GENERATORS_EXPORT setFrequency(float &m_f) Q_DECL_OVERRIDE;
+    bool GENERATORS_EXPORT setAmp(double &m_amp) Q_DECL_OVERRIDE;
+    bool GENERATORS_EXPORT setFrequency(double &m_f) Q_DECL_OVERRIDE;
 
 public:
     bool GENERATORS_EXPORT isGetU(QSerialPortInfo &info);
@@ -23,7 +23,7 @@ public:
 private:
     int timeOutTimerId;
     int poolPeriod;
-    float freq;
+    double freq;
 
 };
